Merge databases for originsrv, sessionsrv and jobsrv into a single builder database
This change comes with a migration script that will need to be run after services are upgraded in tools/merge-databases.

This is the first pass and collapsing our services into a more condensed service that will be easier to manage and deploy.
By merging the databases I have reduced the amount of data duplication across disparate databases and will make it easier to reason about service domains going forward.
I also took the liberty of removing old migrations since we are starting with a pristine database and included a snapshot of the tables and functions we are currently using.

Finally, this removed all of the async code required for package and origin syncing.

Signed-off-by: Elliott Davis <elliott@excellent.io>