Fix to actually use the [memcache] section from config.toml
The config wasn't being picked up from the config.toml due to a
mismatch between the section name in the toml (`memcache`) and the
key in the `Config` structure in `config.rs` (`memcached`).

Thinks were silently working due to the default being localhost:11211.
This showed up on k8s where `sys.ip` for memcached is different from
the host that builder-api is running on.

Added a test case for config.rs and consolidated on `memcache` as the name
that should be the section header.

Signed-off-by: James Casey <james@chef.io>