Clarify the wording around job cancellation
The wording we’re using now is a bit inconsistent; this uses the word “remaining” to include jobs that have either not yet begun or are in progress, and changes the Cancel button to use something other than “cancel”, because the thing you’re confirming happens to be a cancellation.

Also ensures we animate the icon in the jobs list when applicable.

Signed-off-by: Christian Nunciato <chris@nunciato.org>