Consolidate linting
Since we're only running clippy once in the matrix, it makes sense to
only run rustfmt once as well. That will speed up the rest of the matrix
and still give us coverage. We also build habitat-builder-protocol first
now, because the generated files need to be present before rustfmt runs.

Signed-off-by: Josh Black <raskchanky@gmail.com>