Add bootstrap utilities to PATH, instead of binlinking
When running this script on a pre-existing Linux system, we would try
to binlink common utilities that the script uses. However, this would
end up trying to overwrite system utilities like `awk` and `tar`,
which we don't necessarily want to do.

Now, we'll binlink into a temporary directory and add _that_ to the
`PATH` instead. The end result is the same, but we've had a smaller
impact on the system.

Signed-off-by: Christopher Maier <cmaier@chef.io>