Improvments to GitHub web hook handling
This change makes the bldr.toml plan_path more user friendly
by adapting to formats that may not be exactly to spec - for
example, if a user specifies a plan.sh file as part of the plan
path, we can intelligently deal with it.

The change also adds improved logging and other simplifications
by removing parts of the code that are not used/supported.

Signed-off-by: Salim Alam <salam@chef.io>