Programmatically handle enabled features
Rather than managing features on an ad-hoc basis in our Terraform
code, we now can declaratively state which features we want enabled,
and then our templates will do the right thing.

This relies largely on Terraform 0.12's new `templatefile` function,
which allows for a bit of logic within template files, and also allows
non-primitive variable access (in other words, we can use lists, and
not have to create space-delimited strings using `join`).

As a result, this removes the `template_file` resources that were used
for our `hab-sup.service` and `hab-sup.init` files, as well as our
Windows worker user data.

For the Windows worker user data, the caveat is that we can no longer
use `${...}` syntax in the Powershell code to refer to Powershell
variables, because then Terraform thinks it should be templating
those. This is OK, because we can just use `$` without the braces and
everything works out fine.

Also, most of our usage for Powershell was without braces anyway, so
this has the nice side effect of standardizing things. Double win!

Signed-off-by: Christopher Maier <cmaier@chef.io>