Rework main loop to use clap option parsing
The number and variations of the commands has grown substantially. I
got tired of manual parsing so changed it to use clap for the internal
command.  There's a minor fix required to get help output working,
because we can't distinquish a help requirest from a bad arg, but this
already is a much more useful interface IMHO.

Signed-off-by: Mark Anderson <mark@chef.io>