clean up data modeling issues
This PR corrects several issues discovered during preliminary RBAC work:

* adds uniqueness constraints to the origin_members and
origin_invitations tables
* removes duplicate entries from the origin_members table

With these changes, users will not be able to send multiple invitations
to a specific user for any origin. As a result, users should no longer
be able to accept multiple invitations to the same origin which
previously resulted in them having multiple entries for the same origin
in the origins_members table.

As mentioned, the duplicate entries that do currently exist in the origin_members
table will be deleted. We've identified that the calling functions,
including the web UI, handle conflict errors gracefully.

The migrations have been vetted and tested both in local builder studio
and in Acceptance, via transaction and rollback.

Signed-off-by: Jeremy J. Miller <jm@chef.io>