First pass at adding db metrics
This adds metrics to everything that queries the origin_packages table
in a way that looks likely perform poorly.

It follows the existing naming convention as best I could. There's a
lot of boilerplate, which seems a little bit hard to remove.

However I'd like to suggest a different convention: let's construct
the name as structname,functionname using the approprate
inflection/casing scheme. E.g. for the Packages::count_origin_packages
functionthe enum names would be PackagesCountOriginPackagesCallTime,
and the logging keys would be
db-call.packages-count-origin-packages-call-time. I find it easier to
figure out, and I believe it will make it easier to group related
statistics.

This potentially breaks existing logs, and I'd suggest we log twice,
once with the old and once with the new.

One advantage of this scheme is that I believe it makes it easier to
write a macro that would reduce the boilerplate substantially.

Signed-off-by: Mark Anderson <mark@chef.io>