Rename db metrics to use orthogonal scheme
When we only had one or two metrics, the naming scheme made sense, but
as it got extended for more coverage it became somewhat ambiguous.

This renames the metrics to use a more orthogonal scheme, where the
name is composed of structname followed by function name, camel cased
or dash separated as appropriate.

The one downside of this is that it breaks continuity in our logging
system. It might not be a big deal, but if we care we can log the
statistic twice, once with the old name and once with the new.

Signed-off-by: Mark Anderson <mark@chef.io>