Add API to list latest packages for channel
First pass at #1419

This provides a way to list all latest packages for a channel.

We chose the endpoint
/depot/channels/{origin}/{channel}/pkgs/latest?target=TARGET which is
mildly problematic, as it aliases with any package called latest. This
is an existing problem, in that we also alias for promote, but in that
case we can differentiate because one is GET and the other PUT.

One possibility would be to revert to the suggested
/depot/channels/ORIGIN/CHANNEL/latest?target=TARGET endpoint, but that
has it's own problems, as discussed in #1419
(https://github.com/habitat-sh/builder/issues/1419#issuecomment-632667908)

Further testing would be useful; the current test only has one package
in the origin.

Signed-off-by: Mark Anderson <mark@chef.io>