Move API endpoint, and add RAML docs
This moves the API endpoint from '/depot/channels/ORIGIN/CHANNEL/pkgs/latest' to
'/depot/channels/ORIGIN/CHANNEL/latest'. While this is a little irregular, in that it seems to apply
to the channel properties, not the packages in the channel, it avoids accidentally conflicting with
the package name 'latest'.

Also adds a RAML spec for the new endpoint.

Signed-off-by: Mark Anderson <mark@chef.io>