Pin habitat_core, sha2, protobuf(-rust)
Protobuf 2.15 introduces a error (below) in `cargo check`.  Updating to Rust 1.44 no longer errors
when running `cargo check`, so we should re-evaluate the pin when we make that update.

```
error[E0658]: non-builtin inner attributes are unstable
 --> /root/.cargo/registry/src/github.com-1ecc6299db9ec823/protobuf-2.15.0/src/well_known_types/any.rs:9:1
  |
9 | #![rustfmt::skip]
  | ^^^^^^^^^^^^^^^^^
  |
  = note: for more information, see https://github.com/rust-lang/rust/issues/54726
```

sha2 changed their api slightly in the 0.9.0 update. It looks to be a fairly small set of changes we'd need to make
but would have sufficient impact that pinning and splitting the update to minimize churn in a single PR felt like
the right thing to do.

habitat_core is pinned to the previous release of Habitat. A quick inspection of the changes necessary to work the
PackageArchive change through would result in a fairly large diff that merits its own PR, even if it end up being
largely mechanical.

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>