Generate PackageBuildManifest from the acyclic graph.
This adds a trait method compute_build to take a graph and compute an
annotated build graph called the manifest, and extends both cyclic and
acyclic trait implementations to match.

The build manifest represents the transformed graph of packages
rendered buildable by our various algorithms. For the acyclic graph
this is pretty straightforward, but it allows us to use the new
scheduler and the old graph together for testing. The cyclic graph
already implements this process.

This includes test cases for both cyclic and acyclic graphs; they are
similar, but not exactly identical.

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>
Signed-off-by: Mark Anderson <mark@chef.io>
Co-authored-by:	Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>