Notes and design doc
These notes describe our implementation plan for a new job
scheduler. The current scheduler is tightly coupled with the database
representation of the graph, and is unable to handle the
transformations required to handle circular dependencies.

Prior commits introduced the build manifest, which is an abstracted
representation of a build order as transformed to make it tractable.

These notes describe a scheduler that works on a DAG manifest to build
packages.

The DOT files are the best reference for the allowed state transitions.

Signed-off-by: Mark Anderson <mark@chef.io>
Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>
Co-authored-by: Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>