Update worker builds to use $hab_binary
The install path for `hab` in the containers has changed and takes
precedence over the curl|bash installed hab. This change ensure we're
always using the binary we just installed so that we're building using
the latest hab for the correct target.

This change is a sibling commit to https://github.com/habitat-sh/habitat/pull/7854

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>