First pass for new scheduler.
This implements the core scheduler and related subsystems.

This schedules individual jobs within a job group, and handles their
life cycle. This does not handle the processing of groups, selection
of the next group, etc. This also doesn't handle the creation of jobs
from API calls.

Note:
Using project_id requires us to start with the project name, look up
the project record to insert into the job_graph_entry. Later we need
to look up the project to get the name again. It also adds opacity in
the meantime.

Signed-off-by: Mark Anderson <mark@chef.io>
Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>
Co-authored-by:	Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>