Update access token creation for new crypto types
Note that we're removing the `validate` function because it actually
didn't do anything. First, it was simply decrypting data that we had
already decrypted before. Then, it tested to see if the sending and
receiving keys (both of which are actually the same!) needed to
decrypt the message were present on disk, which, of course, they *had*
to be in order to have already decrypted the data!)

Signed-off-by: Christopher Maier <cmaier@chef.io>