Overhaul the integrations module to use new crypto types
We're still assuming that all the encrypting and decrypting is
emitting and receiving base64-encoded strings, even though this isn't
strictly necessary. All encryption and decryption actually works on
data that is already a `String`. Tokens are the only product of
encryption that subsequently need to be base64-encoded, but that can
be handled by itself. We *could* store all the keys in the database
without base64 encoding them, but that would require migrations.

That would also mean that we wouldn't need to return the Builder key
revision used in the encryption process, because the `SecretBox`
returned would be able to tell us that directly.

Signed-off-by: Christopher Maier <cmaier@chef.io>