Decrypt origin secrets using new crypto types
These are anonymous boxes, so we don't actually need a public key to
decode them. As a result, we don't even need to pull them from the
database.

The encrypted secret also doesn't appear to be stored base64-encoded
in the database, so we can just parse it directly as an `AnonymousBox`

Signed-off-by: Christopher Maier <cmaier@chef.io>