Change type from PathBuf to KeyCache
Ideally, we'd also change the name of the struct fields to
`key_cache`. However, that would currently break existing
configuration files, which are deserialized from TOML using the
default `Deserialize` trait implementation.

We could address this with custom `Deserialize` implementations that
are not tied to the internal composition of the structs... that can be
work for another time, though.

Signed-off-by: Christopher Maier <cmaier@chef.io>