Fix race condition in metrics setup
The clone here in an unsafe block created a race condition
    unsafe {
        INIT.call_once(|| {
                SENDER = Box::into_raw(Box::new(init()));
            });
        (*SENDER).clone()
    }

This was causing an increasing number of failures in the metrics code,
including illegal instruction and invalid memory references (SIGILL,
SIGSEV). Our favorite was 'thread panicked while panicking'. While we
don't know exactly what that means, recursive panic sounds
bad. Fortunately the failures happened often enough to be debuggable;
we suspect our increasing usage of tokio async more frequently
triggered this pre-existing issue.

The basic strategy rewrites the block to protect the clone with a
mutex, using the lazy_static macro to eliminate the need for the
unsafe block at all.

One possible risk with this approach is contention over the mutex (or
whatever atomic locking happens inside of lazy_static) causing some
slowdowns in the critical paths of our code.However this is at least
correct, which the old approach was not.

This looks good in local testing, but we will want to run this in
acceptance for a while to ensure that it's a solid fix.

Signed-off-by: Mark Anderson <mark@chef.io>