[MIGRATION] Encrypt origin encryption keys in the database
These secret encryption keys must be encrypted at rest in the
database. This ensures that all paths for storing the keys encrypt
them, and all paths for retrieving them decrypt them before use.

There is some unfortunate duplication in the Job server, but we'll
need to do some broader refactoring before we have a good way of
sharing this logic.

For the migration logic, I'm trying a new approach. Here, it seems we
must use actual Rust code to do the migration, since Postgres doesn't
know anything about our encryption types. I've segregated the
migration code into separate modules in order to prevent mingling of
it with "normal" code. The database logic lives in a
`migration_support` module in the `habitat_builder_db`
crate. Everything needed for looping through all the keys, encrypting
them, and writing them back to the database lives there.

Since the migrations are actually called from the API server on
startup, the logic for invoking the model function in a transaction
lives there in the `server::migrations::encrypt_secret_keys`
module. It's broken out this way in anticipation of future similar
migrations as a way to logically organize things.

Signed-off-by: Christopher Maier <cmaier@chef.io>