Change worker heartbeat log levels
We are occasionally seeing Workers that appear to be in a semi-connected
state. They are listed as ready in our dashboards, work is assigned to
them but the job isn't dispatched. Logging in the acceptance environment
indicates a zmq issue (Failed to dispatch job to worker, err=Zmq(Host unreachable)).

This is a first pass at being able to gather additional information.
Includeing additional state in the logs will help us trace what
machines are involved, along with the work we are attempting to do.
These events should be infrequent and treated as something noteworthy
and worth investigating, so their log level has been adjusted
accordingly.

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>