[REFACTOR] Add new AccessToken type; create Builder token with it
In an attempt to further streamline and clarify code around tokens, a
new `AccessToken` type is introduced. This is the type that is
ultimately responsible for creating the string form of a token that
users interact with.

As an initial step in this refactoring, we'll pull the logic for
creating a Builder token into this new type as a constructor.

Signed-off-by: Christopher Maier <cmaier@chef.io>