[REFACTOR] Pull user token creation logic into new AccessToken type
Also removed an unsightly `unwrap` call. In actuality, this `unwrap`
is probably "fine": the only way it really would have failed was if we
had absolutely no Builder encryption keys present on disk. I'm not at
a point where I can say that would *never* happen, though, so for now,
we'll get rid of the `unwrap` and explicitly handle the error case.

It's a bit unwieldy at the moment, though, given the lack of `?` usage
in this module. That will need some more work.

Signed-off-by: Christopher Maier <cmaier@chef.io>