[REFACTOR] Add AccessToken documentation and unit tests
Now that the refactoring is done and the shape of things is clear, add
documentation on everything. The information flow through the various
types in this part of the code was a little surprising, but hopefully
this refactoring makes it more clear where additional streamlining can
take place.

Adds a bunch of unit tests to valid the behavior and structure of
access tokens, which were completely missing prior to this.

Signed-off-by: Christopher Maier <cmaier@chef.io>