[REFACTOR] Add validation logic to AccessToken type
This allows us to extract a `decrypt` function, as well as a `FromStr`
implementation, further clarifying exactly what is going on in the
process of validating a token, and making it available for potential
reuse elsewhere.

In particular, the `FromStr` implementation checks that:
* the token begins with '_'
* is in base64 encoding
* parses as an encrypted message.

We could make the parsing logic even more strict if we wanted to, but
I'm not sure if it's necessarily worthwhile.

Also, since the `FromStr` implementation incorporates the logic of the
`is_access_token` function, the latter no longer adds any
value. Whenever it was called, we would immediately validate the token
anyway. So we'll just remove it while we're in here.

Signed-off-by: Christopher Maier <cmaier@chef.io>