Extend package build manifest to support existing build workers.
The build workers don't allow exact package resolution, but instead
resolve the latest from a channel. The execution ordering of the base
graph is specified only by the direct dependencies of the
package. This creates an antidependency (read before write) as it is
possible that we build the next iteration of a package before all of
the consumers of the last iteration have started; if that happens
those packages might pick up the wrong iteration. This is likely
harmless, except it makes the process + nondeterministic and hard to
debug. Constraining this will protect against this nondeterminism at
the cost of some parallelism.

Signed-off-by: Mark Anderson <mark@chef.io>
Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>
Co-authored-by: Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>