Plumb the new scheduler into the system, and refine it
This plumbs the new scheduler into the system. It alters the worker
manager to use either the new or old scheduler based on the feature
flag.

It also adds logic to create a job group in handler, create a
manifest, and insert the manifest into database.

This completes the basic end-to-end implmentation.

Signed-off-by: Mark Anderson <mark@chef.io>
Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>