[DB] extend job_graph_mark_complete with as_built info
We need to annotate our graph entries with the final built package
name; this both serves as a bookkeeping feature and the basis of
building a better worker system. In particular, it would be good to be
able to note the expected resolved deps from each worker and verify we
actually used them. We currently rely on some extra ordering in the
graph manifest and the correct behavior of the special build channel
to guarantee this.

Also adds NewJobGraphEntry helper type and fixes up a bunch of lints

Signed-off-by: Mark Anderson <mark@chef.io>
Co-authored-by:	Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>