Removes version from Server: HTTP Response Header
Addresses the low vulnerability reported via CHEF-14845.

This changes the header from "Server: nginx/1.21.4" to "Server: nginx"
which is the best we can do given that we're using the open source
version of nginx.  If we were using the paid version we could eliminate
the head altogether by specifying 'server_tokens "";' in the nginx.conf.

Signed-off-by: Jason Heath <jh@jasonheath.com>