Taking cargo machete to our crates
I found a couple of tools (machete, shears) when I was shrinking the
builder code base that can be use to find unnecessary dependencies in a
rust project. I took another look at those and decided they were fairly
close and chose machete over shears. They seem roughly equivalent and I
think that if we ever wanted to switch it should be rather easy.

Reasons that dependencies were removed include
- The dependency was already available as a transitive dependency
- The shrinking of the code base eliminated the need for the dependency

Serde was added to the ignore list because the parent crate is needed
for the macros. Interestingly shears might be able to detect that in the
future but that feature doesn’t seem to be done baking just yet and I
liked the --with-metadata option in machete

Signed-off-by: Jason Heath <jason.heath@progress.com>