Sets LTS-2024 and guards against glibc mismatches
This commit ensures that LTS-2024 is used everywhere.

This commit also overhauls several scripts in order to deal with glibc
conflicts that led to most of the verify.pipeline failing with segfaults
in the compliation stage. It happens that core/rust/1.79.0 was built
against core/glibc/2.35 while the LTS-2024 packages are built against
core/glibc/2.36. This commit provides a shared build environment that
solves the issue everywhere while drying up several inconsistent
variations of building out the environment.

Signed-off-by: Jason Heath <jason.heath@progress.com>